pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nonebot_plugin_github_release_notifier"
version = "0.1.10"
description = "A plugin for nonebot & onebot to notify group members of new commits, issues, and PRs in GitHub repos."
authors = [
    { name = "HTony03", email = "HTony03@foxmail.com" },
]
readme = "README.md"
keywords = ["nonebot", "github", "qqbot", "reminder"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Plugins",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
license = "MIT"
requires-python = ">=3.10,<=3.13"

dependencies = [
    "aiohttp~=3.11",
    "nonebot-adapter-onebot~=2.4",
    "nonebot-plugin-apscheduler~=0.3",
    "nonebot-plugin-localstore~=0.7",
    "nonebot-plugin-htmlrender~=0.5",
    "nonebot2~=2.3",
    "packaging~=25.0",
    "jinja2~=3.1",
    "markdown-it-py~=3.0",
    "mdit_py_emoji~=0.1",
    "humanize~=4.12",
    "mdit-py-plugins~=0.4",
    "githubkit~=0.13"
]

[project.urls]
homepage = "https://github.com/HTony03/nonebot_plugin_github_release_notifier"
repository = "https://github.com/HTony03/nonebot_plugin_github_release_notifier"
changelog = "https://github.com/HTony03/nonebot_plugin_github_release_notifier/releases"
issue = "https://github.com/HTony03/nonebot_plugin_github_release_notifier/issues"

[dependency-groups]
dev = [
    "nonebot2[fastapi]>=2.4.2,<3.0.0",
    "ruff>=0.11.6,<1.0.0",
    "flake8>=6.0.0,<7.0.0",
    "pydantic>=2.7.0,<3.0.0",
    # "nonemoji>=0.1.4,<1.0.0", # 推荐全局安装 pipx install nonemoji
    # "pre-commit>=4.1.0",      # 推荐全局安装 pipx install pre-commit
]

test = [
    "nonebot2[fastapi]>=2.4.2,<3.0.0",
    "nonebot-adapter-onebot>=2.4.6,<3.0.0",
    "nonebug>=0.3.7,<1.0.0",
    "pytest-xdist>=3.6.1,<4.0.0",
    "pytest-asyncio>=1.0.0,<1.2.0",
    "aiohttp~=3.11,<4.0.0",
]

[tool.nonebot]
plugin_dirs = ["src/"]
adapters = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" }
]


[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
pythonpath = ["src"]
addopts = [
    "-v", # 详细输出
    "-s", # 显示打印信息
    "--tb=short", # 简短的错误回溯
    "-ra", # 显示所有测试结果摘要
    "--strict-markers", # 严格标记模式
    # "--doctest-modules", # 运行文档测试
    "--import-mode=prepend", # 导入模式
]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.format]
line-ending = "lf"

[tool.ruff.lint]
select = [
    "F", # Pyflakes
    "W", # pycodestyle warnings
    "E", # pycodestyle errors
    "I", # isort
    "UP", # pyupgrade
    "ASYNC", # flake8-async
    "C4", # flake8-comprehensions
    "T10", # flake8-debugger
    "T20", # flake8-print
    "PYI", # flake8-pyi
    "PT", # flake8-pytest-style
    "Q", # flake8-quotes
    "TID", # flake8-tidy-imports
    "RUF", # Ruff-specific rules
]
ignore = [
    "E402", # module-import-not-at-top-of-file
    "UP037", # quoted-annotation
    "RUF001", # ambiguous-unicode-character-string
    "RUF002", # ambiguous-unicode-character-docstring
    "RUF003", # ambiguous-unicode-character-comment
    "W191", # indentation contains tabs
    "TID252", # relative import
]


[tool.ruff.lint.isort]
force-sort-within-sections = true
known-first-party = ["tests/*"]
extra-standard-library = ["typing_extensions"]

[tool.ruff.lint.flake8-pytest-style]
fixture-parentheses = false
mark-parentheses = false

[tool.ruff.lint.pyupgrade]
keep-runtime-typing = true


[tool.pyright]
pythonVersion = "3.10"
pythonPlatform = "All"
defineConstant = { PYDANTIC_V2 = true }
executionEnvironments = [
    { root = "./tests", extraPaths = ["./src"] },
    { root = "./src" },
]
typeCheckingMode = "standard"
reportShadowedImports = false
disableBytesTypePromotions = true

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"nonebot_plugin_github_release_notifier" = [
    "lang/*.json",
    "renderer/templates/**/*"
]
